XaVote: Added a "stop votes" menu which stops all current votes.
Made it so you set the short name when you initialize the vote rather than on CreateQuestion as it feels more natural, and it allows us to track the vote instances by short name

Made it so the instances are saved, the only change to the API is the fact CreateVote takes one less argument (the shortname argument).

Added a function to retrieve any global variable (getGlobalVariable(<variable name>) )

xartv:

Now uses xavote to manage its votes

MANY small bugs fixed

the random map generator has been substituted for a better way of randomizing the maps (and more efficient).